requirements.txt
<<<<<<< HEAD
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
altair==4.2.2
amqp==5.2.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
appdirs==1.4.4
async-timeout==4.0.3
asyncssh==2.14.2
atpublic==4.1.0
attrs==23.2.0
AutoTS==0.5.3
billiard==4.2.0
blinker==1.6.2
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.8
cryptography==42.0.8
dacite==1.8.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
dpath==2.2.0
dulwich==0.22.1
dvc==3.51.2
dvc-data==3.15.1
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.20.0
dvc-task==0.4.0
email_validator==2.1.1
entrypoints==0.4
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.1
flatten-dict==0.4.2
flufl.lock==7.1.1
frozendict==2.3.4
frozenlist==1.4.1
fsspec==2024.6.0
funcy==2.0
gitdb==4.0.10
GitPython==3.1.31
grandalf==0.8
gto==1.7.1
h11==0.14.0
html5lib==1.1
htmlmin==0.1.12
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
hydra-core==1.3.2
idna==3.7
ImageHash==4.3.1
iterative-telemetry==0.0.8
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kombu==5.3.7
lightgbm==4.0.0
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
multimethod==1.9.1
multitasking==0.0.11
mysql-connector==2.2.9
mysqlclient==2.2.0
neattext==0.1.3
networkx==3.2.1
numpy==1.26.4
omegaconf==2.3.0
orjson==3.10.4
packaging==24.1
pandas==2.2.2
pandas-profiling==3.6.6
pathspec==0.12.1
pconst==1.0.0
phik==0.12.3
platformdirs==3.11.0
prompt_toolkit==3.0.47
psutil==5.9.8
pyarrow==11.0.0
pycparser==2.22
pydantic==2.7.3
pydantic-extra-types==2.8.0
pydantic-settings==2.3.1
pydantic_core==2.18.4
pydeck==0.8.1b0
pydot==2.0.0
pygit2==1.15.0
Pygments==2.15.1
pygtrie==2.5.0
Pympler==1.0.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pywin32==306
PyYAML==6.0.1
requests==2.32.3
rich==13.3.4
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-learn==1.5.0
scipy==1.13.1
scmrepo==3.3.5
semver==3.0.2
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.1
sqltrie==0.11.0
starlette==0.37.2
streamlit==1.21.0
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
threadpoolctl==3.5.0
tomlkit==0.12.5
tqdm==4.66.4
typeguard==2.13.3
typer==0.12.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==4.3
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.1
validators==0.20.0
vine==5.1.0
visions==0.7.5
voluptuous==0.14.2
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
wordcloud==1.9.2
xgboost==1.7.6
yarl==1.9.4
ydata-profiling==4.5.1
yfinance==0.2.9
zc.lockfile==3.0.post1
=======
﻿aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
altair==4.2.2
amqp==5.2.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
appdirs==1.4.4
async-timeout==4.0.3
asyncssh==2.14.2
atpublic==4.1.0
attrs==23.2.0
AutoTS
billiard==4.2.0
blinker==1.6.2
celery==5.4.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.8
cryptography==42.0.8
dacite==1.8.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dnspython==2.6.1
dpath==2.2.0
dulwich==0.22.1
dvc==3.51.2
dvc-data==3.15.1
dvc-http==2.32.0
dvc-objects==5.1.0
dvc-render==1.0.2
dvc-studio-client==0.20.0
dvc-task==0.4.0
email_validator==2.1.1
entrypoints==0.4
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.1
flatten-dict==0.4.2
flufl.lock==7.1.1
frozendict==2.3.4
frozenlist==1.4.1
fsspec==2024.6.0
funcy==2.0
gitdb==4.0.10
GitPython==3.1.31
grandalf==0.8
gto==1.7.1
h11==0.14.0
html5lib==1.1
htmlmin==0.1.12
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
hydra-core==1.3.2
idna==3.7
ImageHash==4.3.1
iterative-telemetry==0.0.8
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kombu==5.3.7
lightgbm==4.0.0
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
multimethod==1.9.1
multitasking==0.0.11
mysql-connector==2.2.9
mysqlclient==2.2.0
neattext==0.1.3
networkx==3.2.1
numpy==1.26.4
omegaconf==2.3.0
orjson==3.10.4
packaging==24.1
pandas==2.2.2
pandas-profiling==3.6.6
pathspec==0.12.1
pconst==1.0.0
phik==0.12.3
platformdirs==3.11.0
prompt_toolkit==3.0.47
psutil==5.9.8
pyarrow==11.0.0
pycparser==2.22
pydantic==2.7.3
pydantic-extra-types==2.8.0
pydantic-settings==2.3.1
pydantic_core==2.18.4
pydeck==0.8.1b0
pydot==2.0.0
pygit2==1.15.0
Pygments==2.15.1
pygtrie==2.5.0
Pympler==1.0.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pywin32==306
PyYAML==6.0.1
requests==2.32.3
rich==13.3.4
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
scikit-learn==1.5.0
scipy==1.13.1
scmrepo==3.3.5
semver==3.0.2
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.16.0
smmap==5.0.0
sniffio==1.3.1
sqltrie==0.11.0
starlette==0.37.2
streamlit==1.21.0
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
threadpoolctl==3.5.0
tomlkit==0.12.5
tqdm==4.66.4
typeguard==2.13.3
typer==0.12.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==4.3
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.1
validators==0.20.0
vine==5.1.0
visions==0.7.5
voluptuous==0.14.2
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
wordcloud==1.9.2
xgboost==1.7.6
yarl==1.9.4
ydata-profiling==4.5.1
yfinance==0.2.9
zc.lockfile==3.0.post1

>>>>>>> c036d56c5c5692131f2007a3466a2bd4902034dc
